requirements.txt
asgiref==3.7.2
cloudinary==1.33.0
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.19
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
django-filter==23.2
django-multiselectfield>=0.1.12
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
oauthlib==3.2.2
Pillow==8.2.0
psycopg2==2.9.6
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.16
whitenoise==6.5.0
